fdo#74573: Skip deletion of destination area when 'skip empty' is on.
Also, adjust handling of mix document aka paste functions with this
change.  When using paste function (add, subtract, etc), the behavior
between the 'skip empty' flag on and off makes no difference.  Let's
set the flag to off when paste function is used.

+++ sc/source/core/data/document.cxx
@@ -2690,7 +2690,8 @@ void ScDocument::CopyFromClip( const ScRange& rDestRange, const ScMarkData& rMar
         SCCOL nCol2 = pRange->aEnd.Col();
         SCROW nRow2 = pRange->aEnd.Row();
 
-        DeleteArea(nCol1, nRow1, nCol2, nRow2, rMark, nDelFlag);
+        if (!bSkipAttrForEmpty)
+            DeleteArea(nCol1, nRow1, nCol2, nRow2, rMark, nDelFlag);
 
         if (CopyOneCellFromClip(aCxt, nCol1, nRow1, nCol2, nRow2))
             continue;
